refactor/#6 [rename] : comm 디렉터리를 util, work 디렉터리로 분리 & home 디렉터리 삭제
