+++ app/src/main/java/com/devnuts/ruflu/domain/repository/ChatRepository.kt
@@ -1,7 +1,7 @@
 package com.devnuts.ruflu.domain.repository
 
-import com.devnuts.ruflu.comm.retrofit.RufluApp
-import com.devnuts.ruflu.comm.retrofit.ServerAPI
+import com.devnuts.ruflu.util.RufluApp
+import com.devnuts.ruflu.util.ServerAPI
 import com.devnuts.ruflu.ui.model.chat.ChatRoom
 import retrofit2.Call
