+++ app/src/main/java/com/devnuts/ruflu/ui/adapter/NearByAdapter.kt
@@ -7,7 +7,7 @@ import android.widget.ImageView
 import androidx.lifecycle.MutableLiveData
 import androidx.recyclerview.widget.RecyclerView
 import com.devnuts.ruflu.R
-import com.devnuts.ruflu.comm.utill.UserUtil
+import com.devnuts.ruflu.util.UserUtil
 import com.devnuts.ruflu.ui.model.home.UserDtl
 
 class NearByAdapter(val data: MutableLiveData<ArrayList<UserDtl>>) : RecyclerView.Adapter<NearByAdapter.NearByViewHolder>() {
