+++ app/src/main/java/com/devnuts/ruflu/ui/adapter/RufluLikeLv1Adapter.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.devnuts.ruflu.R
-import com.devnuts.ruflu.comm.utill.UserUtil
+import com.devnuts.ruflu.util.UserUtil
 import com.devnuts.ruflu.databinding.LikeLv1UserBinding
 import com.devnuts.ruflu.ui.like.fragment.RufluTouchHelperCallback
 import com.devnuts.ruflu.ui.model.home.UserDtl
