+++ app/src/main/java/com/devnuts/ruflu/ui/adapter/UserCardViewAdapter.kt
@@ -10,7 +10,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.ViewPager2
 import com.devnuts.ruflu.R
-import com.devnuts.ruflu.comm.utill.UserUtil
+import com.devnuts.ruflu.util.UserUtil
 import com.devnuts.ruflu.ui.home.fragment.HomeSubSEFrag
 import com.devnuts.ruflu.ui.home.viewmodel.HomeSubSEViewModel
 import com.devnuts.ruflu.ui.model.home.UserCard
