+++ app/src/main/java/com/devnuts/ruflu/util/BaseDiffUtil.kt
@@ -1,4 +1,4 @@
-package com.devnuts.ruflu.comm.utill
+package com.devnuts.ruflu.util
 
 import androidx.recyclerview.widget.DiffUtil
