+++ app/src/main/java/com/devnuts/ruflu/worker/AppNotification.kt
@@ -1,16 +1,14 @@
-package com.devnuts.ruflu.comm
+package com.devnuts.ruflu.worker
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.os.Build
-import android.util.Log
 import androidx.annotation.RequiresApi
 import androidx.core.app.NotificationCompat
 import com.devnuts.ruflu.R
-import com.devnuts.ruflu.comm.retrofit.RufluApp
-import okhttp3.internal.notify
+import com.devnuts.ruflu.util.RufluApp
 import timber.log.Timber
 
 class AppNotification(rufluApp: RufluApp) {
