+++ app/src/main/java/com/devnuts/ruflu/worker/CustomCardStackView.kt
@@ -1,4 +1,4 @@
-package com.devnuts.ruflu.home
+package com.devnuts.ruflu.worker
 
 import android.content.Context
 import android.util.AttributeSet
