+++ app/src/main/java/com/devnuts/ruflu/worker/FusedLocationProvider.kt
@@ -1,10 +1,9 @@
-package com.devnuts.ruflu.comm
+package com.devnuts.ruflu.worker
 
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Looper
-import android.util.Log
 import androidx.core.app.ActivityCompat
 import com.devnuts.ruflu.MainActivity
 import com.google.android.gms.location.*
