+++ app/src/main/java/com/devnuts/ruflu/worker/LogInterceptor.kt
@@ -1,4 +1,4 @@
-package com.devnuts.ruflu.comm
+package com.devnuts.ruflu.worker
 
 import java.text.SimpleDateFormat
 import okhttp3.Interceptor
