+++ app/src/main/java/com/devnuts/ruflu/worker/ReceivedCookiesInterceptor.kt
@@ -1,6 +1,6 @@
-package com.devnuts.ruflu.comm
+package com.devnuts.ruflu.worker
 
-import com.devnuts.ruflu.comm.retrofit.RufluApp
+import com.devnuts.ruflu.util.RufluApp
 import okhttp3.Interceptor
 import okhttp3.Response
